Move arithmetic functions to a separate submodule
And add docs for implemented functions
