+++ src/builtin/functions/mod.rs
@@ -50,12 +50,14 @@ macro_rules! intern_set_func {
     };
 }
 
+mod arithmetic_operations;
 mod conditionals;
 mod functions;
 mod list_elements;
 mod sequences;
 
 pub(crate) fn add(ctx: &mut TulispContext) {
+    arithmetic_operations::add(ctx);
     conditionals::add(ctx);
     functions::add(ctx);
     sequences::add(ctx);
